Capstone_Project-CS[02]/Codebase/requirements.txt
# CV Sorting using LLMs (CS[02]) — Python Dependencies
# Install with:
#   pip install -r requirements.txt

# LLM API client (Groq)
groq

# Resume parsing
pypdf          # PDF parsing
python-docx    # Word (.docx) parsing

# PDF report generation
fpdf2
